Create navbar component and CSS
Navbar created to with close button. Close handler passed in as a prop from header component.
Menu items will be changed to the appropriate routes instead of the anchor tags they are now.
CSS input to style close button and menu items. Colours will be changed to fit colour scheme